Fixed vertical overflow
Swish text is now height-constrained, scrollable, and height scales to window size.

Added a three-step scaling gradient based on vertical resolution to fix the overflow coming back on extremely high- or low-resolution displays, and narrowed the line spacing very slightly to fix the overflow on the third slide on the top row which previously happened due to the added subtitle.