Resolution scaling and slight visual tweak - FIX
Added a three-step scaling gradient based on vertical resolution to fix the overflow coming back on extremely high- or low-resolution displays, and narrowed the line spacing very slightly to fix the overflow on the third slide on the top row which previously happened due to the added subtitle.

Edited - accidentally left some visual changes I was meaning to undo in the last commit.