Updated method of code preloading to external file and changes to button and formatting, cleaned up and added comments
Code is now preloaded into Swish by inputting the address of the external file hosted on GitHub in the COMS30106/prolog_scripts repository. No changes need to be made to the existing examples present or to the method in which new examples are added, as the address is created from the ID of the <pre> tag, which is already present for all of the Swish boxes. A new button and element-swapping system had to be implemented to facilitate this, but the visual change is minimal. Currently, the contents of the box before Swish is activated are still pulled from the 'local' set of prolog scripts, however, so these should not be removed.

Changed the button to display in a better location which will also work for the vertical-scrolling book format. Reverted potentially problematic changes to line-height. Changed the way the button is inserted into the document such that modifications to the swish_box.html file are no longer necessary and have been undone.

Cleaned up the code, consolidated all of the $(document).ready(function(){}) statements into one, removed extraneous code, and added comments.