Updated method of code preloading to external file
Code is now preloaded into Swish by inputting the address of the external file hosted on GitHub in the COMS30106/prolog_scripts repository. No changes need to be made to the existing examples present or to the method in which new examples are added, as the address is created from the ID of the <pre> tag, which is already present for all of the Swish boxes. A new button and element-swapping system had to be implemented to facilitate this, but the visual change is minimal. Currently, the contents of the box before Swish is activated are still pulled from the 'local' set of prolog scripts, however, so these should not be removed.