Changes to button and formatting
Changed the button to display in a better location which will also work for the vertical-scrolling book format. Reverted potentially problematic changes to line-height. Changed the way the button is inserted into the document such that modifications to the swish_box.html file are no longer necessary and have been undone.